// src/shared/api/mold-passport/model/castingTechnologyLookupsListResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Molding Shop Service
 * OpenAPI spec version: 0.1.0
 */
import type { CastingTechnologyLookupResponse } from './castingTechnologyLookupResponse'

export interface CastingTechnologyLookupsListResponse {
  data: CastingTechnologyLookupResponse[]
  page: number
  totalPages: number
  totalItems: number
}