// src/shared/api/mold-passport/model/moldingFlaskLookupsListResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Molding Shop Service
 * OpenAPI spec version: 0.1.0
 */
import type { MoldingFlaskLookupResponse } from './moldingFlaskLookupResponse'

export interface MoldingFlaskLookupsListResponse {
  data: MoldingFlaskLookupResponse[]
  page: number
  totalPages: number
  totalItems: number
}