// src/shared/api/mold-passport/model/patternPlateFrameLookupsListResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Molding Shop Service
 * OpenAPI spec version: 0.1.0
 */
import type { PatternPlateFrameLookupResponse } from './patternPlateFrameLookupResponse'

export interface PatternPlateFrameLookupsListResponse {
  data: PatternPlateFrameLookupResponse[]
  page: number
  totalPages: number
  totalItems: number
}